requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
crawl4ai>=0.4.0
exa_py>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
httpx>=0.27.0
